requirements.txt
Flask==0.12.2
gunicorn==19.8.1
Werkzeug==0.12.2
psycopg2-binary==2.7.7
python-dotenv==0.10.1
cython==0.29.6
pandas==0.23.0
numpy==1.16.1
requests2==2.16.0
flask-request-params==0.3.0
flask-cors==3.0.7
sqlalchemy==1.3.7
GeoAlchemy2==0.6.3
uuid==1.30